validaciones en alumnos y docentes , y arreglo del index para mostrar correctamente las cards

+++ src/main/java/ar/edu/unju/fi/util/ListaAlumnos.java
@@ -13,9 +13,9 @@ public class ListaAlumnos {
 		
 		listaAlumnos = new ArrayList<Alumno>();
 		
-		Alumno alumno1 = new Alumno(1234, "facu", "xfs", "facuxfs@gmail.com", (float) 87887.0);
-		Alumno alumno2 = new Alumno(7824, "sol", "xfs", "solxfs@gmail.com", (float) 78787.0);
-		Alumno alumno3 = new Alumno(1004, "ro", "gump", "rogump@gmail.com", (float) 77701.0);
+		Alumno alumno1 = new Alumno(1234, "facu", "xfs", "facuxfs@gmail.com",  87887);
+		Alumno alumno2 = new Alumno(7824, "sol", "xfs", "solxfs@gmail.com",  78787);
+		Alumno alumno3 = new Alumno(1004, "ro", "gump", "rogump@gmail.com",  77701);
 		
 		listaAlumnos.add(alumno1);
 		listaAlumnos.add(alumno2);
